fix: move settings override normalization out of server actions module
